Define the `research` command structure in `rmplan.ts` steps 1-5
## [1mStep 1[22m

In `src/rmplan/rmplan.ts`, add a new command definition for `research [planArg]`. This should follow the pattern of other commands in the file, like `next` or `done`.


## [1mStep 2[22m

Set the description for the `research` command to "Generate a research prompt for a plan and append the results."


## [1mStep 3[22m

Add a boolean option `--rmfilter` to the command, with the description "Use rmfilter to include file context in the prompt."


## [1mStep 4[22m

Configure the command to allow excess arguments and unknown options, which will be used to pass file paths to `rmfilter`. Use `.allowExcessArguments(true)` and `.allowUnknownOption(true)`.


## [1mStep 5[22m

In the `.action()` handler for the `research` command, dynamically import `handleResearchCommand` from a new file `src/rmplan/commands/research.ts` and call it, passing the `planArg`, `options`, and `command` arguments. Ensure this is wrapped in a `.catch(handleCommandError)` block.