Implement the paste-back and plan update functionality steps 1-5
## [1mStep 1[22m

In `src/rmplan/actions.ts`, modify the `handleResearch` function. After copying the prompt, log a message instructing the user on the next steps: perform the research, and then paste the results back into the terminal. Use `sshAwarePasteAction` to provide the correct instruction for their environment.


## [1mStep 2[22m

Use `waitForEnter(true)` from `src/common/terminal.ts` to capture the multi-line pasted content from the user.


## [1mStep 3[22m

If the pasted content is not empty, append it to the `details` field of the plan object. Prepend the new content with a clear separator like `\n\n--- Research ---\n\n` to distinguish it from the original details.


## [1mStep 4[22m

Update the `updatedAt` timestamp on the plan object to the current time.


## [1mStep 5[22m

Use `writePlanFile` to save the modified plan object back to its file, overwriting the previous version. Log a confirmation message that the plan was updated.