Update `writePlanFile` to save plans in the new front matter format steps 1-2
## [1mStep 1[22m

In `src/rmplan/plans.test.ts`, add a test case for `writePlanFile`. It should create a `PlanSchema` object that includes a `details` field. After calling `writePlanFile`, read the content of the newly created file. Assert that the file is formatted with YAML front matter (the `details` field should not be present in the YAML part) and that the original `details` string constitutes the body of the file.


## [1mStep 2[22m

In `src/rmplan/plans.ts`, update the `writePlanFile` function. It should separate the `details` field from the rest of the plan object. Serialize the plan object (without `details`) into a YAML string. Construct the final file content by wrapping the YAML string in `---` separators, followed by the `details` content. Ensure the existing `yaml-language-server` schema comment is included within the front matter block.